Created specClassNamesForSim1CT, to ensure that only the correct 3 are used.
Using f.specChoiceOneCTCleaned() throughout. Uses reactive().  NOT assigned to an output component.

You DO need "isolate", to prevent f.specChoiceOneCTCleaned() from running when f.specChoiceOneCTCleaned() is run. Only a change in the numericInput should trigger it.

This is close. The indices do stick.

BUGS TODO:The "0" shows up, and the specChoiceOneCT pops back to PopModelSpecifier.